Add an explicit exception to LGDT/LIDT.
A #UD exception is thrown when a non-memory operand is referenced. This is not
explcitly required by the latest version of the Intel manuals, but it is by an
earlier version. Comments to that effect have been added.

Thanks to Shilpi Goel for finding the assertion in the older version of the
manual.