VL: fix obscure bug with introducing enum constants
Enums are tricky due to how they introduce constants.  In other
Verilog tools, it is not permitted to write:

   enum { FOO, BAR } a;   // <-- introduces FOO and BAR
   enum { FOO, BAR } b;   // <-- introduces FOO and BAR again, name clash error

But it *is* permitted to write:

   enum { FOO, BAR } a, b;

VL did not support this because we split up declarations of the latter
form right away, leading us to think this was a name clash.

Fix this by tweaking the enumnames transform to remove duplicates.
Add some test cases to validate the fix.