Book changes needed for change to evaluators.
Most changes were due to renaming of constraint theorems -- the ones
formerly named constraint-N for N >= 6 are now named constraint-N+1.  A
few books needed the new constraint (constraint-6, or
-of-nonsymbol-atom) to be enabled in order to show a functional
instantiation correct.  The clause processor utility ev-find-rules.lisp
needed a new function to find the nonsymbol-atom rule.  A couple of
books contained expansions of defevaluator forms, and we modified these
to match the new defevaluator output.  Finally, tools/defevaluator-fast
used to contain a second implementation of defevaluator, but the
built-in defevaluator had then been changed to support its same set of
features, so we changed defevaluator-fast to just call the built-in
defevaluator and removed the other implementation.