[x86isa] Remove vex-prefixes and evex-prefixes args from some functions
One-, two-, and three-byte opcode dispatch functions are not called by
vex-decode-and-execute or evex-decode-and-execute anymore; as such,
the former don't need to have vex-prefixes and evex-prefixes as
arguments.

If the presence of vex or evex prefixes is illegal, it will be
detected by vex-decode-and-execute or evex-decode-and-execute now, and
not inside the instruction semantic functions.