GL updates:
- Reworked the way GL selects rewrite rules to try.  Formerly, we went
through all the ACL2 rewrite rules for the leading function symbol,
skipping those that weren't listed as GL rules.  Now, instead, we use a
memoized function to extract only the list of eligible rules from the
world.  Similar for branch merge rules (except that the function that
computes the whole table is memoized, because all rules are hung on IF
in the ACL2 world).

- Pass the global GL interp-st into GL/GLMC clause processors instead of
creating a local one.  This supports querying profiling info from
aborted runs.