VL: clocking blocks and "default disable" parsing
Fix confusion in the parsing code for clocking blocks.  We had been using
backtracking to handle cases where the clocking_event begins with an
identifier, but that is impossible: they always start with @.  Also fix an
oversight in the element parsing code: it was only invoking the clocking block
parser for "clocking", but not for "default clocking."

Add a parser and representation for "default disable" blocks.  Note that like
clocking blocks, these are effectively ignored throughout most of VL and are
not supported in any sensible way, but at least we will now be able to parse
them.

Extend linttest/parse to check that we can parse a module with a default
clocking block and a default disable construct.