Improved :doc on lisp programmer intro.  Improved :args for IF.  Improved a couple of error messages.  Strengthened table guard for user-defined-functions-table.  Explained defattach non-redundancy in a comment.  Added missing :doc parent.
Improved :doc
introduction-to-programming-in-acl2-for-those-who-know-lisp and added
defpointer, lisp-programmer-introduction, to it.

Quoting :doc note-8-1:

  Fixed :[args] to avoid hard ACL2 error when applied to IF and to
  provide a clearer error message for Common Lisp functions not in
  ACL2.  Thanks to Eric McCarthy for sending examples to point out
  these issues.

Tweaked verify-guards error message to mention "function symbol"
before "theorem name", as suggested by Eric Smith.

Quoting a comment in (defxdoc note-8-1 ...):

; Restricted user-defined-functions-table to require the correct number of
; formals in each value function.

Added a comment in the #+acl2-loop-only definition of defattach for
why defattach events are not redundant, and pointed to that comment in
:doc redundant-events.  Thanks to Eric Smith for raising the question
of redundancy for defattach.

Added missing parent to :doc smt::fncall-of-flextype (and got the OK
on this from Yan Peng).