Improve JMP far.
Instead of reading a 16-byte descriptor right away, first read an 8-byte
descriptor, because code segment descriptors are always 8 bytes, also in 64-bit
mode. Then check if it is a valid code segment descriptor: if it is, process
accordingly. Otherwise, read a 16-byte descriptor, check if it is a call gate,
and if so process it accordinglty. The code refactoring done in the previous
commit has enabled the change made by this commit.

Also fix the check of the segment selector index against the table limit: the
index must be multiplied by 8 and increased by 7 (when reading the 8-byte
descriptor) and 15 (when reading the 16-byte descriptor), before being compared
with the table limit.