Finish extending JMP far to application-level 32-bit mode.
In the application-level view, only code segments are used with JMP far. System
segments are used in the system-level view, which is currently not supported in
32-bit mode anyhow (it needs the paging kinds used in protected mode).

Since the handling of code segment in JMP far has been extended to 32-bit mode
in previous commits, I've now moved the check for 64-bit mode after code segment
processing. Thus, JMP far can be executed both in 64-bit and 32-bit mode with
code segments now; it can be still only executed in 64-bit mode with system
segments. The handling of system segments by JMP far will be extended to 32-bit
mode eventually.

I also had to change the sign (via I64) of the offset read from memory when the
operand size is 8 bytes, because addresses are signed in 64-bit mode in our
model.

This commit marks an important milestone in the X86ISA model: all the
non-floating-point, application-level instructions have been extended to 32-bit
mode. This coincides with the status described in the final version of the paper
for the upcoming ACL2 Workshop.

The floating-point instructions, and JMP far with system segments (described
above) will be extended to 32-bit mode eventually, and so will paging. At that
point, also 32-bit system programs will be supported.