[x86isa] ea-to-la returns a canonical linear address
This affects 64-bit mode only.

Previously, in the 64-bit mode, ea-to-la would perform the canonicity
check only if seg-reg was FS or GS, which meant that for these
seg-regs, the canonicity check was unnecessarily done again in the
rme*/wme* functions.  Now, ea-to-la always returns a canonical linear
address upon a successful translation, which means that rme*/wme*
functions don't have to perform this check.

Also: added some documentation for rip-guard-okp.