Eliminated a hard error for congruence rules in the case of local defequiv events.  Fixed mod to books/GNUmakefile from earlier today to clean centaur/getopt/demo2.dx86cl64 always.
Quoting :doc note-8-2:

  Eliminated a hard error labeled as ``Implementation error'' that
  could occur when submitting a :[congruence] rule during the second
  pass of [encapsulate] or the local incompatibility check in Step 3
  of [certify-book].  The error occurred when the that the
  equivalence relation of the rule had been defined locally, hence
  was missing during that second pass or local incompatibility check.
  Now, a useful ordinary 9``soft'') error occurs, with a useful
  message.  Thanks to Nathan Guermond for reporting this bug.

Thanks to Eric Smith for noticing that cleaning of
centaur/getopt/demo2.dx86cl64 was for Mac only.  I found the existing
clean of centaur/getopt/demo2.lx86cl64 and put the new one next to
that, for books always to be cleaned rather than only with target
"moreclean".