Add new DEFMAX-NAT macro.
This is a macro to introduce a declarative definition of the form

  f(x1,...,xn) = max {y in Nat | psi<x1,...,xn,y>}

where Nat is the set of natural numbers, psi<x1,...,xn,y> is a formula, and f
returns NIL if the set has no maximum. The user supplies psi, x1,...,xn, y, an
optional guard for f (t by default), and an optional flag (t by default) to
indicate whether guards should be verified.

The macro is invoked as follows:

(defmax-nat y (x1 ... xn)
  psi<x1,...,xn,y>
  :guard guard<x1,...,xn>
  :verify-guards t/nil)

(Note that the syntax for the bound variable y is analogous to defchoose.)

The macro generates a (non-executable) definition of f, along with some
auxiliary functions and several theorems. It also generates two theorems to help
reason about f: one theorem serves to establish that a maximum exists (so f is
not NIL) if we can exhibit a member of the set and an upper bound of the set;
another theorem serves to establish that the maximum has a certain value, if we
can show that such a value is a member of the set and is an upper bound of the
set.

This macro should be easy to generalize to integers, and to other suitably
ordered domains. Besides maxima, similar macros could be added to declaratively
specify minima, infima, and suprema.