Added partial-encapsulate macro, to avoid needing a ttag for unknown-constraints.
Quoting :doc note-8-2:

  A new macro, [partial-encapsulate], allows one to introduce
  constrained functions without specifying all of the [constraint]s.
  This functionality was already available using a trust tag, by way
  of a rather convoluted application of
  [define-trusted-clause-processor]; however, [partial-encapsulate]
  may be used without a trust tag.  See [partial-encapsulate], which
  in particular points to an example of typical usage, in
  books/demos/partial-encapsulate.lisp.  Thanks to Sol Swords for
  requesting such a capability.

There were quite a few implementation changes in support of this new
macro, which should be invisible to users.  However, the change to
trusted-clause-processor-table will be visible to anyone who accesses
it explicitly: it has a different shape, and hence has been renamed to
trusted-cl-proc-table in order to avoid silent errors caused by this
shape change.