Changed pso (etc.) to report the Time from the original summary.  Added a user-settable maximum (default 9) for depth of nested inductions.  Fixed :doc in certify-book and include-book regarding guard-checking.
Quoting :doc note-8-2:

  For calls of :[pso] and related utilities (:[pso!], :[psof], and
  :[psog]), the Time reported in the summary is now the original
  time, not time related to running :[pso].

  There is now, by default, a limit of 9 on the nesting depth of
  inductions; see [induction-depth-limit] and to modify this default,
  see [set-induction-depth-limit].

The induction-depth-limit idea is based on a discussion I had some
time ago with J.  This change may make it a bit less likely that
beginners get put off ACL2 due to proofs going into the weeds.

Fixed the :doc for certify-book and include-book to say that
guard-checking is on during these operations.  (It was formerly off,
was on starting with Version 7.2.)  Thanks to Alessandro Coglio for
pointing out this issue.