Add a formalization of Ethereum's Modified Merkle Patricia trees.
This is a complex data structure, which combines characteristics of Merkle trees
and Patricia (a.k.a. radix) trees, with a modification that optimizes sub-paths
of single-child nodes. MMP trees use cryprographic hashes as
implementation-independent "pointers" (supported by a database used an
implementation-independent "heap"). These hashes are visible at the Ethereum
interface, as root hashes. Thus, MMP trees are not merely implementation-level
entities, but they belong, in suitable form, to the high-level formal
specification of Ethereum as well. The mathematical possibility of hash
collisions is explicitly taken into account in this formalization, but we do not
expect to ever see a :COLLISION result in practice, due to Keccak-256's
collision resistance.