Add a test generation facility to ATJ.
A new input :TESTS is evaluated and the result is checked to be a list of
doublets (name term), where name is a string and term is an untranslated term
that must translate to a call of the form (fn qc1 qc2 ...), where fn is one of
the target functions of ATJ and qc1, qc2, etc. are quoted constants.

For each such test, ATJ generates Java code to test the call as follows:
- Build the AIJ representations of the values of qc1, qc2, etc.
- Evaluate the call in AIJ, obtaining a result value.
- Compare the result value with the one resulting from the evaluation of the
  call in ACL2 (ATJ performs this evaluation, and generates Java code to build
  this value so that it can be compared with the result of the call in AIJ.
- Print a message saying whether the test has passed or not.

This supports the fast creation of many tests just by writing ground calls, and
letting ACL2 calculate the correct result, to be tested against AIJ's result.

The testing code is generated in a separate class (see documentation). If :TESTS
is NIL, no testing class is generated. The testing class, when generated, has a
main() method that runs all the test, and can be invoked as a Java application
from a shell.