[X86ISA] Remove now-obsolete operand access functions.
The functions X86-OPERAND-FROM-MODR/M-AND-SIB-BYTES, X86-OPERAND-TO-REG/MEM, and
X86-OPERAND-TO-XMM/MEM have now been removed. They have been replaced by
functions with the same names but with an ending $. These $ functions have a
slightly different interface than the non-$ function, necessitated by the
extension of the model to 32-bit mode. While the non-$ functions were adequate
to 64-bit mode, the $ functions are adequate to both 64-bit and 32-bit mode.

After introducing the $ functions, calls to the non-$ functions got gradually
replaced by calls to the $ functions, as instructions were being extended to
32-bit mode. Now that all the instructions have been extended to 32-bit mode,
the non-$ functions are no longer used, and therefore have been eliminated in
this commit.

The next commit will rename the $ functions to remove the $. It seemed best to
carry out all of this in two commits (this one and the next) rather than in one.