[X86ISA] Improve the DEFTHM-USB utility.
This macro is analogous to DEFTHM-NATP, but for UNSIGNED-BYTE-P instead of NATP.

The default hints have been improved to be the most minimal to prove the
corollaries from the main theorem. However, since the user can supply different
hyps for the corollaries, the user still has the ability to supply explicit
hints for the corollaries. But if the hyps for the corollaries trivially subsume
the hyps for the main theorem, then the automatically generated hints should
suffice.