[x86isa] WIP: Opcode maps are now defined using defprod
The old opcode map lists were obtained more-or-less automatically by
parsing the Intel manuals.  Though they corresponded pretty nicely to
the relevant pages of the manual, it was confusing to work with them;
for instance, one had to constantly check annoying things like the
following: doing an nth of a simple cell inside a compound cell
results in a true-listp.

We obtained the new product-type data structures for opcode maps from
the old lists.  With this change, the opcode maps are organized more
uniformly, and as such, it's easier to write tools based off them.
Moreover, they're still quite readable: it's easy to see where this
information comes from in the Intel manuals.

TODO: fix dispatch utils, etc. to read off the new maps.