Tweaked defun-sk, primarily by making it sensitive to the default-verify-guards-eagerness and delaying guard verification to the end.
Quoting :doc note-8-2:

  [Defun-sk] is now sensitive to the [default-verify-guards-eagerness],
  and guard verification is always delayed to near the end of the
  generated event to avoid failures due to the small theory present
  at defun time.  Thanks to Alessandro Coglio for emails on leading
  to these improvements.  Some additional small tweaks have been made
  to defun-sk, in particular to check that there are not two or more
  distinct values associated with [xargs] keywords :verify-guards,
  :non-executable, or (even if not distinct) :guard-hints.

Added books/system/tests/defun-sk-tests.lisp to check some of the
changes to defun-sk.

Note that when :constrain t is supplied, then :guard and
:verify-guards continue to be essentially ignored; only :guard is
relevant, and only in the local definition.  On completion of the
defun-sk, the guard will always be t for :constrain t and the
resulting constrained function will be considered to be
guard-verified.

Made mods to :doc std::define-sk and to
books/kestrel/utilities/define-sk-tests.lisp to reflect that guard
verification in defun-sk is now always delayed, so no longer subject
to certain failures due to small theories.