[X86ISA] Fix bug in segment register selection.
The function SELECT-SEGMENT-REGISTER uses, as default segment register
(i.e. when no segment override prefix is present), either DS or SS, based on the
16-bit and 32-bit addressing modes in Tables 2-1 and 2-2 of Intel's Volume 3.
However, the function missed a case in which SS must be the default: the case is
when Table 2-2 requires a SIB byte, and the Base field of the SIB byte specifies
rSP as the base register. This has been now fixed. It was necessary to add the
SIB byte as an argument of SELECT-SEGMENT-REGISTER, so numerous calls of this
function had to be slightly modified to pass the SIB byte.

Thanks to Eric Smith for uncovering an anomaly in the course of symbolic
execution, and for helping track the issue to the bug described above.