[X86ISA] Handle the DS/ES/FS/GS/SS.W bit.
The W bit of a data segment descriptor says whether the data segment is
read-only (W = 0) or read/write (W = 1). In 32-bit mode, when writing data, we
must ensure that W is 1 and that the segmnent is not a code segment. These
checks are only performed in 32-bit mode.

So the W(I)MExx and W(I)ME-SIZE functions have been extended to include this
check.

Also adapted the 32-bit Fibonacci execution, by initializing the state so that
the stack segment has SS.W = 1.