[X86ISA] Add checks for null segment selectors.
When accessing memory through the DS, ES, FS, and GS segment registers, these
registers cannot contain a null segment selector (which can be nonetheless
loaded into them without error). A check to this effect has been added to the
EA-TO-LA function, which translates effective address to linear addresses.

A similar check is not needed for CS and SS, because attempting to load a null
segment selector into those registers causes an exception. In other words, it is
a state invariant that CS and SS never contain a null segment selector.

Adapted a couple of theorems to this change.