[ETHEREUM] Improve the RLP parser and decoders.
Return more detailed error information as first result, instead of just T for
error and NIL for no error.

Revise the correctness theorems of the executable decoders to say that the first
result is IFF-equivalent to the declaratively defined decoders (while the second
result is still EQUAL-equivalent).

Add some tests for the new kinds of errors returnes.