Use lock-free hash table in CCL for fast-alists.  Fixed a partial-encapsulate test to work with GCL.  Avoid error in building manual when zip isn't installed.  Improved :doc trace$.
Quoting :doc note-8-2:

  (CCL only) We now use lock-free hash tables for [fast-alists], to
  work around an apparently CCL bug.  Thanks to Rob Sumners and the
  folks at Centaur for finding and analyzing this problem, proposing
  this fix, and doing timing tests on it.

Quoting :doc note-8-2-books:

  We now avoid causing an error when building the manual in the case
  that the zip program is not installed; that causes a warning
  instead.

Thanks to Michał Herda for feedback leading to the enhancement above.
(It's conventional to add credits in ACL2 release notes but not in
book release notes.)  I decided to remove the dependency on zip rather
than to state on the installation pages that zip is required.

Improved :doc trace$ by pointing out that relevant variables are in
the ACL2 package, and by improving formatting a bit.