[FTY] Add a macro to define fixtype "aliases".
The FTY::DEFFIXTYPE-ALIAS macro introduces an alias for an existing fixtype, and
optionally its recognizer, fixer, and equivalence. It uses ADD-MACRO-FN for the
functions, and FTY::DEFFIXTYPE for the "alias" fixtype.

Thanks to Sol Swords for suggesting that FTY::DEFFIXTYPE can be used to create
aliases in this sense, since it merely records an association between a fixtype
name and its recognizer, fixer, and equivalence. This new macro
FTY::DEFFIXTYPE-ALIAS is just a convenience that creates, besides the new
fixtype, also aliases of recognizer, fixer, and equivalence.

This new macro could be a candidate for inclusion into the FTY library.