[JAVA] Optimized AIJ's evaluation of primitive functions.
The Java representation of ACL2 primitive functions has been improved and
generalized to a notion of 'native' functions (i.e. implemented natively in
Java), vs. 'defined' functions that have an ACL2 definition in the environment
and that are executed via interpretation by AIJ.

The native functions are represented by individual classes and singleton
instances that override the function application method. In this way, the
previous naive case analysis on the name of each invoked primitive function is
now replaced with Java's dynamic dispath.

The times of the current tests before and after this change are actually not too
different. Probably the reason is that the number of primitive function calls
are not a significant fraction of the total number of calls.