[CRYPTO] Collect the cryptographic placeholders.
The cryptographic placeholders in the Bitcoin and Ethereum libraries have been
collected into a more central placeholder for a cryptographic library.

The Bitcoin and Ethereum library now contain just thin wrappers of these
collected placeholders.

In particular, this ensures that the Bitcoin and Ethereum libraries agree on the
cryptographic functions (as opposed to having different constrained functions as
respective placeholders), which facilitates the creation of artifacts that use
components from the two libraries, such as wallets.