[XDOC] Improve XDOC constructors.
Use a newly introduced type of trees to represent structured XDOC strings, which
are eventually converted to strings by the top-level XDOC constructor. This
enforces a type discipline on the construction of XDOC strings, and opens the
door to more optimized ways to performed the final string flattening than by
simple string concatenation as done now.

Also add more XDOC constructors for description lists.

Remove the XDOC constructor tests, since they are just simple "unit tests" and
we have a lot of real uses of XDOC constructors to test them.