Take a suggestion by @ericwhitmansmith
A note for Eric: I'm really surprised that your suggestion of removing
the hypotheses worked! I'd run remove-hyps on both these rules before
making this pull request, and no hypotheses were removed in either
case. This might be something to talk to Matt and Nathan - the original
authors of remove-hyps - about.