[X86ISA] Add extended version of SELECT-OPERAND-SIZE.
SELECT-OPERAND-SIZE is used by many instructions, but several instructions do
not use it because they follow slightly different rules, for determining operand
size, than the "common" rules. As a result, these other instructions all include
the same block of code to calculate the operand size.

The new version of SELECT-OPERAND-SIZE includes additional parameters that
accommodate the slightly different rules for operand size determination. All the
instructions should now be able to use this new version, reducing code
duplication and improving clarity and modularity.

In order to stage the change, since the new function has a different interface (
more parameters) than the old function, the new function is currently called
SELECT-OPERAND-SIZE$. Existing calls to SELECT-OPERAND-SIZE will be redirected
to SELECT-OPERAND-SIZE$, which will also be called by the instructions that
currently do not call SELECT-OPERAND-SIZE. When SELECT-OPERAND-SIZE is no longer
called, it will be removed, and SELECT-OPERAND-SIZE$ will be renamed to
SELECT-OPERAND-SIZE.