Add an integer arithmetic library.
This is mainly the work of Nathan Guermond.

I had to verify the guards of memb in [books]/sorting/perm.lisp to avoid a
conflict with [books]/meta/term-defuns.lisp in the manual. Perhaps the new
integer library is somehow switching the order in which these two defuns of memb
are processed.