(1) Avoid undesirable HIDE calls (caused by scons-term).  (2) Fixed ev-fncall-rec-logical to comprehend the :attach nil argument of defattach.  (3) Made a few other tweaks.
Details:

.....

(1) The fix for undesirable HIDE calls caused by scons-term is in
response to the following example from J.  The problem was that when a
call of ev-fncall caused an error when made on behalf of scons-term,
the result was always placed inside a call of HIDE.  That is now
avoided if the failure was due to the need for a warrant hypothesis,
since that can be dealt with later by the rewriter.  (The trick is to
call ev-fncall+ instead of ev-fncall.)

(include-book "projects/apply/top" :dir :system)

(defun$ sum (fn lst)
    (declare (xargs :guard (and (apply$-guard fn '(nil))
                                (true-listp lst))))
    (if (endp lst)
        0
        (+ (fix (apply$ fn (list (car lst))))
           (sum fn (cdr lst)))))

(defun$ sq (x)
    (declare (xargs :guard (integerp x)))
    (* x x))

(defun$ nats (n)
    (if (zp n)
        nil
        (cons n (nats (- n 1)))))

(include-book "arithmetic-5/top" :dir :system)

; This formerly failed because a HIDE was introduced, but now
; it succeeds.
(thm
 (implies (and (warrant sq)
               (natp n))
          (equal (sum 'sq
                      (nats n))
                 (/ (* n (+ n 1) (+ (* 2 n) 1))
                    6))))

.....

(2) There was no easy way to look at the world and decide whether an
attachment was made with argument :attach nil of defattach.  The new
world global, attach-nil-lst, is a list of function symbols attached
using :attach nil.  Thanks to J with helping to design this solution.

.....

(3) Other tweaks included: other minor changes to scons-term; fixes to
the partial-encapsulate introducing ev-fncall+-fns to deal more
appropriately with the strictp argument: and (c) trivial variable name
improvements in expand-abbreviations.