[OMAP & FTY] Add a new macro FTY::DEFOMAP.
This is analogous to FTY::DEFLIST (for lists), FTY::DEFALIST (for alists), and
FTY::DEFSET (for osets): it generates a fixtype for types omaps whose keys and
values have specified fixtypes. It also generates a number of theorems,
similarly to the other macros.

Note that FTY::DEFMAP is already taken by an existing macro that generates
alists with a specific fixing function. Maybe names could be made more
consistent at some point (including the current "dissonance" between the package
names SET and OMAP, as opposed to the more consistent options (i) SET and MAP or
(ii) OSET and OMAP).