[Java] Add a preliminary formalization of primitive values.
These are tagged values, which enable (in the future) the definition of a
defensive semantics and a proof of type soundness, as often done programming
language formalizations.

The model of floating-point values is abstract for now, with constrained
recognizers that are placeholders for a future precise model.

The formalization includes (abstract) extended-exponent values, based on
parameters (constrained nullary functions) that specify whether these values are
supported and in what way.