Removed item from developer to-do list: "Consider redefining take using mbe, so that the :logic has a standard recursion."
Recent commits completed that task, but as noted in my preceding
commit, I forgot to mention the change to take's built-in ACL2
definition in the message for the commit that made that change.  Here
is the actual release note.

Quoting :doc note-8-2:

  The built-in function [take] now has a recursive definition, exactly
  along the lines of the theorem take-redefinition from the community
  book books/std/lists/take.lisp (written by Jared Davis), which is
  retained for compatibility with existing books.  The definition of
  take uses [mbe], where the :exec component calls first-n-ac as
  before for execution efficiency.  We thank Mihir Mehta for
  providing this enhancement, including updates to the books.

The item I removed from to-do.txt mentioned the following; however, I
think it's not worth bothering with, since the optimization is
specifically for fixnums and could, I suppose, benefit some future
Lisp implementation even if it doesn't benefit one today.

> A more ambitious project would be to try to remove the
> #-acl2-loop-only code in take in favor of :exec code in an mbe.  This
> would include experiments in at least CCL and SBCL to check
> performance.