Update developers' guide
This clears up a part of the developers' guide that I found a little
hard to follow. It also, while keeping the end result the same,
slightly changes the order of operations in which an integrator merges
a contributor's branch, in order to reduce the number of merge commits
produced.