Clarified about risk of certifying as root.
I added the following to :doc books-certification.

  It may be preferable to avoid being logged in as root, since
  developers do not test as root and at least one community book
  (books/oslib/tests/copy.lisp) has failed to certify when logged in
  as root.

I also made brief mention of this issue in the installation
instructions.

If someone wants to log in as root and fix the book above, that would
be great.  In that case the paragraph above could be modified
(weakened) accordingly.  But something should remain unless there is
regular testing using root.

Thanks to folks on the acl2-devel list who helped me to think this
through.