[Std/alists] Add some theorems about REMOVE-ASSOC-EQUAL.
The function is disabled at the end of the file, after the theorems are proved.

Adjust an existing proof, which apparently relies on the function to be enabled
after including Std/alists, by enabling the function in the hints. Perhaps this
can be avoided by adding more theorems to the newly added Std/alists file on
REMOVE-ASSOC-EQUAL.