[JAVA] Extend AIJ's public interface.
Some of AIJ's code has been extended and refactored to provide public access (in
the Java sense) to the native Java implementations of the ACL2 primitive
functions.

This is in preparation for an upcoming extension of ATJ with the ability to
generate Java code from ACL2 according to a shallow embedding approach (as
opposed to the current deep embedding approach). In the shallow embedding
approach, the generated Java code must be able to call, eventually, the Java
native implementations of the ACL2 primitive functions, by calling suitable
public methods in AIJ. This commit adds those methods.