[Java] Extend ATJ to support a shallow embedding approach.
A new input, :DEEP (T or NIL), is now available to choose between a deep and a
shallow embedding approach to Java code generation from ACL2.

The deep embedding approach is the one described in the ACL2 Workshop paper,
where we generate Java code that builds deeply embedded Java representations of
the ACL2 functions and execute them via AIJ, the interpreter of ACL2 written in
Java. This approach is simple but has the inherent overhead of interpretation.

The shallow embedding approach is new. Every ACL2 function is translated into a
Java method with the same number of arguments, and whose body mimics the
computations of the ACL2 function by being a shallowly embedded representation
of the ACL2 function body. This is a more complex translation, but avoids
interpretation altogether and is therefore more efficient and more
idiomatic. Preliminary results on non-trivial examples (in particular, the
verified ABNF parser) suggest an order of magnitude speed improvement over the
deep embedding.

Even though the shallow embedding approach does not use AIJ's interpreter, it
still uses AIJ's representation of ACL2 values and native Java implementation of
the ACL2 primitive functions.

Both deep and shallow embedding still execute ACL2 functions (in Java) "in the
logic", i.e. ignoring guards. Supporting ACL2's guard-checking modes of
execution is future work, and should not be too difficult. Running
guard-verified code will also enable further optimizations, such as the use of
native Java types (e.g. primitive integers) when guards suitably restrict
function inputs to be representable with the more restricted types. (Currently
this is not possible, because execution "in the logic" must handle whichever
ACL2 value is thrown at it.)

In the process of adding this shallow embedding option, the structure of the ATJ
code has been improved. For instance, there are now function names with 'deep'
and 'shallow' in their names to reflect their roles.

The documentation has been updated. Tests for the shallow embedding will be
added soon to the repo.