[crypto] Temporarily exclude the attachments from the manual.
When these are included in the manual, there are some conflicts in the
manual. These do not arise from the attachments themselves, but from libraries
indirectly included by the attachments, via some of the recently added
cryptographic libraries.

The attachments will be re-included into the manual when those conflicts are
resolved.