Improved efficiency of xdoc string processing.
Quoting :doc note-8-3-books:

  Some [xdoc] string processing has been made a bit more efficient.

The change was to replace remove-equal-with-hint by a new function,
remove1-name, in books/xdoc/top.lisp.  The primary change was to
remove only the first occurrence of a given :doc topic; I'm assuming
that this is usually the only occurrence but, more important, is the
only occurrence that matters.  Two less important changes are that
during the search for the topic to remove, only the name is compared,
and that cons is now used instead of cons-with-hint (since the cdr
will never match with eq).

In preliminary experiments this cut more than 4% of the time from
(include-book "centaur/sv/top" :dir :system).  I plan to get more
accurate timing information on linux, at which time I may update the
release note.