[Java] Improve Java code generated by ATJ.
Now the code generator avoids creating new Java local variables for lambda-bound
variables that are applied to themselves, i.e. to the extra variables that
normally do not appear in LETs but that ACL2 generates internally in order to
keep lambda expressions closed.

This makes the generated code more readable, but it does not seem to make it
particularly faster, presumably because the Java JIT optimizes assignments of
the form x = y with x and y variables (which is somewhat reassuring that the JIT
is doing a good job).