[Java] Improve the Java code generated by ATJ.
Since ACL2 symbols typically have uppercase letters, we flip those to lowercase
when generating Java variable and method names corresponding to ACL2 variable
and function names; we also flip lowercase letters to uppercase in order to
maintain "injectivity" (no conflicts). We do the flipping for the SYMBOL-NAME
part of the symbol. For the SYMBOL-PACKAGE-NAME part, we do the flipping if the
package name is used as a disambiguating prefix of a Java variable, but we keep
the letters uppercase when generating Java class names corresponding to ACL2
package names.

Overall, this scheme results in more readable and idiomatic Java code.

In the process, fixed a small bug that consisted in improperly testing generated
method names against Java keywords.