[Java] Extend ATJ to generate narrower types.
When :DEEP is NIL (i.e. in the shallow embedding approach) and :GUARDS is T
(i.e. the generated Java code can assume that all guards are satisfied), now ATJ
has the ability to generate Java methods (corresponding to ACL2 functions) with
narrower Java types than the class type Acl2Value of all ACL2 values.

This relies on recording the input and output types of the ACL2 functions of
interest in the ACL2 world, via a macro DEF-ATJ-FUNCTION-TYPE that is now
provided by ATJ. This macro attempts to prove theorems asserting that a
function's guard ensures certain specified types for the arguments and result of
the function (this can be extended to MV results later, but for now it's only
one result per function, MV being treated as a list). If the theorems are
successful, the type information is recorded in an ACL2 table.

Based on the information in this table, ATJ generates Java methods with argument
and result types consistent with the ones of the corresponding ACL2 functions:
in general, these may be narrower than Acl2Value, e.g. if a function has a guard
STRINGP for an argument and the ATJ type :STRING is specified for that argument
in a DEF-ATJ-FUNCTION-TYPE for the function, then ATJ will generate a Java
method where that argument has type Acl2String.

While translating ACL2 terms into Java expressions and statements, ATJ
recursively calculates types for such terms, in order to insert, where needed,
casts to narrower types. In general, there may be a relatively complicated
reason/proof why the argument of a function turns out to be always, say, a
string, and so we need these Java type casts, but they may be rare in
well-structured/well-typed code.

This extension paves the way to have ATJ generate Java code that uses other
types than AIJ's classes, in particular generate Java code that uses primitive
types (and operators), arrays, and so on. This will require an extension of the
currenly simple type system used by ATJ, but the general mechanisms should be in
place now.